# ts/packages/web/src/api/openapi.json
{"openapi":"3.0.2","info":{"title":"Fast API","version":"0.1.0"},"paths":{"/api/res/{resid}":{"get":{"summary":"Getres","operationId":"getres_api_res__resid__get","parameters":[{"required":true,"schema":{"title":"Resid","type":"string"},"name":"resid","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Intent"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Putres","operationId":"putres_api_res__resid__put","parameters":[{"required":true,"schema":{"title":"Resid","type":"string"},"name":"resid","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Intent"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/res":{"get":{"summary":"Findres","operationId":"findres_api_res_get","parameters":[{"required":false,"schema":{"title":"Type","type":"string"},"name":"type","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Intent"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/res/{resid}/train":{"post":{"summary":"Train","operationId":"train_api_res__resid__train_post","parameters":[{"required":true,"schema":{"title":"Resid","type":"string"},"name":"resid","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Content":{"title":"Content","required":["name","examples"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"examples":{"title":"Examples","type":"array","items":{"type":"string"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Intent":{"title":"Intent","required":["type","content"],"type":"object","properties":{"type":{"title":"Type","type":"string"},"content":{"$ref":"#/components/schemas/Content"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}